fix: require an integer context window override
A context window is a token count, and the sibling tier threshold
already validates as z.number().int().min(0). Without .int() a
fractional override like 20000.5 passed validation and reached the
compaction arithmetic.

+++ src/config/schemas-model.ts
@@ -37,7 +37,7 @@ const ModelDefSchema = z.object({
   // field Zod strips it silently at config load (the exact failure mode
   // #1847 shipped for pricing.tiers), and the override never reaches
   // src/agents/native/models.ts#resolveContextWindow.
-  contextWindow: z.number().positive().optional(),
+  contextWindow: z.number().int().positive().optional(),
   env: z.record(z.string(), z.string()).optional(),
 });
